Moved the project package under the legacy package.
