+++ agents_playground/__init__.py
@@ -4,9 +4,9 @@ from .app import *
 from .core import *
 from .entities import *
 from .funcs import *
+from .legacy import *
 from .navigation import *
 from .paths import *
-from .project import *
 from .renderers import *
 from .legacy import *
 from .simulation import *
