+++ agents_playground/legacy/project/rules/scene_file_exists.py
@@ -1,6 +1,6 @@
 import os
-from agents_playground.project.project_loader_error import ProjectLoaderError
-from agents_playground.project.rules.project_validator import ProjectValidator
+from agents_playground.legacy.project.project_loader_error import ProjectLoaderError
+from agents_playground.legacy.project.rules.project_validator import ProjectValidator
 
 class SceneFileExist(ProjectValidator):
   def validate(self, module_name: str, module_path: str) -> bool:
