+++ agents_playground/templates/base_files/scene.py
@@ -4,7 +4,7 @@ from typing import cast, Generator, Tuple
 import dearpygui.dearpygui as dpg
 
 from agents_playground.agents.spec.agent_spec import AgentLike
-from agents_playground.project.extensions import (
+from agents_playground.legacy.project.extensions import (
   register_entity, 
   register_renderer, 
   register_task
