+++ demo/landscape_spike/landscape_spike/scene.py
@@ -2,7 +2,7 @@ from math import copysign, radians
 from typing import cast, Generator, Tuple
 
 from agents_playground.agents.spec.agent_spec import AgentLike
-from agents_playground.project.extensions import (
+from agents_playground.legacy.project.extensions import (
   register_entity, 
   register_renderer, 
   register_task
