+++ demo/line_of_sight/line_of_sight/scene.py
@@ -18,7 +18,7 @@ from agents_playground.core.constants import DEFAULT_FONT_SIZE
 from agents_playground.core.task_scheduler import ScheduleTraps
 from agents_playground.counter.counter import Counter, CounterBuilder
 from agents_playground.fp.containers import FPList
-from agents_playground.project.extensions import register_entity, register_renderer, register_task
+from agents_playground.legacy.project.extensions import register_entity, register_renderer, register_task
 from agents_playground.renderers.color import BasicColors, ColorUtilities, Colors
 from agents_playground.legacy.scene.scene import Scene
 from agents_playground.simulation.context import SimulationContext, Size
