+++ demo/paths/paths/scene.py
@@ -7,7 +7,7 @@ from agents_playground.agents.spec.agent_spec import AgentLike
 from agents_playground.core.task_scheduler import ScheduleTraps
 from agents_playground.paths.linear_path import LinearPath
 from agents_playground.paths.circular_path import CirclePath
-from agents_playground.project.extensions import register_renderer, register_task
+from agents_playground.legacy.project.extensions import register_renderer, register_task
 from agents_playground.renderers.color import Color, PrimaryColors
 from agents_playground.legacy.scene.scene import Scene
 from agents_playground.simulation.context import SimulationContext, Size
